Admin Signout API and Admin Dashboard(Drawer)

+++ admin-app/src/actions/auth.actions.js
@@ -1,5 +1,5 @@
 import {authConstants} from "./constants";
-import axiosInstance from '../helpers/axios';
+import axios from '../helpers/axios';
 
 export const login = (user) => {
   console.log(user);
@@ -9,7 +9,7 @@ export const login = (user) => {
     dispatch({
       type : authConstants.LOGIN_REQUEST
     })
-    const res = await axiosInstance.post('/admin/signin' , {
+    const res = await axios.post('/admin/signin' , {
       ...user
     } )
 
@@ -73,10 +73,28 @@ export const isUserLoggedIn = () => {
 
 export const signout = () => {
   return async dispatch => {
-    localStorage.clear();
+
+    dispatch({
+      type : authConstants.LOGOUT_REQUEST
+    })
+
+    const res=await axios.post('/admin/signout')
+
+    if(res.status == 200){
+      localStorage.clear();
       dispatch({
-        type : authConstants.LOGOUT_REQUEST
+        type : authConstants.LOGOUT_SUCCESS
       });
+    }else{
+      dispatch({
+        type : authConstants.LOGOUT_FAILURE,
+        payload : {
+          error : res.data.error
+        }
+      });
+    }
+
+    
       
   }
 }
